Fix thread safety, fd lifecycle, and security issues across TCP servers and clients
- IQ server: non-blocking broadcast (MSG_DONTWAIT) prevents pipeline freeze on slow clients;
  atomic add_client_fd eliminates TOCTOU race and fd leaks; inet_ntop replaces inet_ntoa
- IQ client: fd_mutex protects sock_fd between recv thread and stop to prevent double-close;
  sample_rate made atomic; header sample_rate=0 rejected; recv_exact handles partial reads
  across timeouts and EINTR
- CAT server: stop uses shutdown() only, handler threads own close() to prevent double-close;
  atomic add_client_fd with count check; send() with MSG_NOSIGNAL and EINTR handling
- CAT control/client: write/send retry on EINTR; response parsing scans for ';' in full
  buffer instead of only checking last byte
- FFT processor: uint32_t casts in sample conversion to avoid undefined behavior from
  shifting into sign bit
- Waterfall: bin index clamped to [0, size-1] to prevent OOB; int64_t casts in bandwidth
  bin calculations to prevent overflow
- Main: demod_last_update changed to atomic_int_least64_t for safe reads on 32-bit ARM (Pi)
- USB device: guard against effective_rate <= 0 from bad EEPROM correction; skip transfer
  free if callbacks still pending to avoid use-after-free
- Settings: validate loaded ref/range values to prevent display glitches from corrupted config
- Build: add -fstack-protector-strong, -Wformat-security; _FORTIFY_SOURCE=2 for release builds

+++ src/fft_processor.c
@@ -114,7 +114,8 @@ void fft_processor_free(fft_processor_t *fft) {
 
 // Convert 32-bit signed integer (4 bytes, little-endian) to float
 static inline float convert_32bit_sample(const uint8_t *data) {
-    int32_t value = data[0] | (data[1] << 8) | (data[2] << 16) | (data[3] << 24);
+    int32_t value = (int32_t)((uint32_t)data[0] | ((uint32_t)data[1] << 8) |
+                              ((uint32_t)data[2] << 16) | ((uint32_t)data[3] << 24));
     // Normalize to [-1.0, 1.0]
     return (float)value / 2147483648.0f;
 }
